rename interactive-sdk to interactive
