+++ coordinator/gscoordinator/flex/core/interactive/hqps.py
@@ -24,13 +24,13 @@ import time
 from typing import List
 from typing import Union
 
-import interactive_sdk.openapi
+import interactive.openapi
 import requests
-from interactive_sdk.openapi import CreateGraphRequest
-from interactive_sdk.openapi import CreateProcedureRequest
-from interactive_sdk.openapi import SchemaMapping
-from interactive_sdk.openapi import StartServiceRequest
-from interactive_sdk.openapi import UpdateProcedureRequest
+from interactive.openapi import CreateGraphRequest
+from interactive.openapi import CreateProcedureRequest
+from interactive.openapi import SchemaMapping
+from interactive.openapi import StartServiceRequest
+from interactive.openapi import UpdateProcedureRequest
 
 from gscoordinator.flex.core.config import CLUSTER_TYPE
 from gscoordinator.flex.core.config import HQPS_ADMIN_SERVICE_PORT
@@ -85,10 +85,10 @@ class HQPSClient(object):
                     return f"http://127.0.0.1:{HQPS_ADMIN_SERVICE_PORT}"
 
     def list_graphs(self) -> List[dict]:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceGraphManagementApi(
+            api_instance = interactive.AdminServiceGraphManagementApi(
                 api_client
             )
             graphs = [g.to_dict() for g in api_instance.list_graphs()]
@@ -115,10 +115,10 @@ class HQPSClient(object):
             return graphs
 
     def get_schema_by_id(self, graph_id: str) -> dict:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceGraphManagementApi(
+            api_instance = interactive.AdminServiceGraphManagementApi(
                 api_client
             )
             schema = api_instance.get_schema(graph_id).to_dict()
@@ -129,30 +129,30 @@ class HQPSClient(object):
             return schema
 
     def create_graph(self, graph: dict) -> dict:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceGraphManagementApi(
+            api_instance = interactive.AdminServiceGraphManagementApi(
                 api_client
             )
             response = api_instance.create_graph(CreateGraphRequest.from_dict(graph))
             return response.to_dict()
 
     def delete_graph_by_id(self, graph_id: str) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceGraphManagementApi(
+            api_instance = interactive.AdminServiceGraphManagementApi(
                 api_client
             )
             rlt = api_instance.delete_graph(graph_id)
             return rlt
 
     def get_graph_by_id(self, graph_id: str) -> dict:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceGraphManagementApi(
+            api_instance = interactive.AdminServiceGraphManagementApi(
                 api_client
             )
             g = api_instance.get_graph(graph_id).to_dict()
@@ -196,10 +196,10 @@ class HQPSClient(object):
         raise RuntimeError("Create vertex type is not supported yet!")
 
     def create_stored_procedure(self, graph_id: str, stored_procedure: dict) -> dict:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceProcedureManagementApi(
+            api_instance = interactive.AdminServiceProcedureManagementApi(
                 api_client
             )
             response = api_instance.create_procedure(
@@ -210,11 +210,11 @@ class HQPSClient(object):
             return response
 
     def list_stored_procedures(self, graph_id: str) -> List[dict]:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
             stored_procedures = []
-            api_instance = interactive_sdk.openapi.AdminServiceProcedureManagementApi(
+            api_instance = interactive.AdminServiceProcedureManagementApi(
                 api_client
             )
             stored_procedures = [
@@ -225,10 +225,10 @@ class HQPSClient(object):
     def update_stored_procedure_by_id(
         self, graph_id: str, stored_procedure_id: str, stored_procedure: dict
     ) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceProcedureManagementApi(
+            api_instance = interactive.AdminServiceProcedureManagementApi(
                 api_client
             )
             return api_instance.update_procedure(
@@ -240,10 +240,10 @@ class HQPSClient(object):
     def delete_stored_procedure_by_id(
         self, graph_id: str, stored_procedure_id: str
     ) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceProcedureManagementApi(
+            api_instance = interactive.AdminServiceProcedureManagementApi(
                 api_client
             )
             return api_instance.delete_procedure(graph_id, stored_procedure_id)
@@ -251,10 +251,10 @@ class HQPSClient(object):
     def get_stored_procedure_by_id(
         self, graph_id: str, stored_procedure_id: str
     ) -> dict:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceProcedureManagementApi(
+            api_instance = interactive.AdminServiceProcedureManagementApi(
                 api_client
             )
             return api_instance.get_procedure(graph_id, stored_procedure_id).to_dict()
@@ -263,10 +263,10 @@ class HQPSClient(object):
         # get service status from serving graph
         serving_graph_id = None
         rlts = []
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceServiceManagementApi(
+            api_instance = interactive.AdminServiceServiceManagementApi(
                 api_client
             )
             response = api_instance.get_service_status()
@@ -304,37 +304,37 @@ class HQPSClient(object):
         return rlts
 
     def stop_service(self) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceServiceManagementApi(
+            api_instance = interactive.AdminServiceServiceManagementApi(
                 api_client
             )
             return api_instance.stop_service()
 
     def restart_service(self) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceServiceManagementApi(
+            api_instance = interactive.AdminServiceServiceManagementApi(
                 api_client
             )
             return api_instance.restart_service()
 
     def start_service(self, graph_id: str) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceServiceManagementApi(
+            api_instance = interactive.AdminServiceServiceManagementApi(
                 api_client
             )
             return api_instance.start_service(StartServiceRequest(graph_id=graph_id))
 
     def list_jobs(self) -> List[dict]:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceJobManagementApi(
+            api_instance = interactive.AdminServiceJobManagementApi(
                 api_client
             )
             rlt = []
@@ -351,10 +351,10 @@ class HQPSClient(object):
             return rlt
 
     def get_job_by_id(self, job_id: str) -> dict:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceJobManagementApi(
+            api_instance = interactive.AdminServiceJobManagementApi(
                 api_client
             )
             job_status = api_instance.get_job_by_id(job_id).to_dict()
@@ -368,10 +368,10 @@ class HQPSClient(object):
             return job_status
 
     def delete_job_by_id(self, job_id: str) -> str:
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceJobManagementApi(
+            api_instance = interactive.AdminServiceJobManagementApi(
                 api_client
             )
             return api_instance.delete_job_by_id(job_id)
@@ -402,10 +402,10 @@ class HQPSClient(object):
         self._job_config[graph_id] = config
         self.dump_to_disk()
         # submit
-        with interactive_sdk.openapi.ApiClient(
-            interactive_sdk.openapi.Configuration(self._hqps_endpoint)
+        with interactive.ApiClient(
+            interactive.Configuration(self._hqps_endpoint)
         ) as api_client:
-            api_instance = interactive_sdk.openapi.AdminServiceGraphManagementApi(
+            api_instance = interactive.AdminServiceGraphManagementApi(
                 api_client
             )
             response = api_instance.create_dataloading_job(
