+++ flex/interactive/sdk/examples/java/interactive-example/pom.xml
@@ -2,7 +2,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.alibaba.graphscope</groupId>
-  <artifactId>interactive-sdk-example</artifactId>
+  <artifactId>interactive-example</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
   <build>
@@ -17,12 +17,12 @@
       </plugin>
     </plugins>
   </build>
-  <name>interactive-sdk-example</name>
+  <name>interactive-example</name>
   <url>http://maven.apache.org</url>
   <dependencies>
     <dependency>
       <groupId>com.alibaba.graphscope</groupId>
-      <artifactId>interactive-sdk</artifactId>
+      <artifactId>interactive</artifactId>
       <version>0.3</version>
     </dependency>
   </dependencies>
