+++ flex/interactive/sdk/generate_sdk.sh
@@ -16,10 +16,10 @@
 # This script is used to generate the Java SDK from the Flex Interactive API
 # It uses the Swagger Codegen tool to generate the SDK
 
-PACKAGE_NAME="com.alibaba.graphscope.interactive.openapi"
+PACKAGE_NAME="com.alibaba.graphscope.interactive"
 PYTHON_PACKAGE_NAME="openapi"
 GROUP_ID="com.alibaba.graphscope"
-ARTIFACT_ID="interactive-sdk"
+ARTIFACT_ID="interactive"
 ARTIFACT_URL="https://github.com/alibaba/GraphScope/tree/main/flex/interactive"
 VERSION="0.3"
 EMAIL="graphscope@alibaba-inc.com"
@@ -57,7 +57,7 @@ function do_gen_java() {
 
     cmd="openapi-generator-cli generate -i ${OPENAPI_SPEC_PATH} -g java -o ${OUTPUT_PATH}"
     cmd=" ${cmd} --api-package ${PACKAGE_NAME}.api"
-    cmd=" ${cmd} --model-package ${PACKAGE_NAME}.model"
+    cmd=" ${cmd} --model-package ${PACKAGE_NAME}.models"
     cmd=" ${cmd} --invoker-package ${PACKAGE_NAME}"
     cmd=" ${cmd} --package-name ${PACKAGE_NAME}"
     cmd=" ${cmd} --artifact-id ${ARTIFACT_ID}"
@@ -72,8 +72,8 @@ function do_gen_python() {
     echo "Generating Python SDK"
     OUTPUT_PATH="${CUR_DIR}/python"
     cmd="openapi-generator-cli generate -i ${OPENAPI_SPEC_PATH} -g python -o ${OUTPUT_PATH}"
-    cmd=" ${cmd} --invoker-package ${PYTHON_PACKAGE_NAME}"
-    cmd=" ${cmd} --package-name interactive_sdk.openapi"
+    # cmd=" ${cmd} --invoker-package ${PYTHON_PACKAGE_NAME}"
+    cmd=" ${cmd} --package-name interactive"
     cmd=" ${cmd} --additional-properties=packageVersion=${VERSION},pythonVersion=3"
     echo "Running command: ${cmd}"
     eval $cmd
