+++ flex/interactive/sdk/java/src/main/java/com/alibaba/graphscope/interactive/client/common/Status.java
@@ -15,8 +15,8 @@
  */
 package com.alibaba.graphscope.interactive.client.common;
 
-import com.alibaba.graphscope.interactive.openapi.ApiException;
-import com.alibaba.graphscope.interactive.openapi.ApiResponse;
+import com.alibaba.graphscope.interactive.ApiException;
+import com.alibaba.graphscope.interactive.ApiResponse;
 
 /**
  * Mapping http status code to our status code, along with a message
