+++ flex/interactive/sdk/python/.openapi-generator-ignore
@@ -25,8 +25,8 @@
 
 README.md
 setup.py
-interactive_sdk/__init__.py
-interactive_sdk/client
+interactive/__init__.py
+interactive/client
 requirements.txt
 test-requirements.txt
 test/
