--- flex/interactive/sdk/python/interactive_sdk/client/__init__.py
@@ -1,6 +1,0 @@
-# flake8: noqa
-
-from interactive_sdk.client.driver import Driver
-from interactive_sdk.client.result import Result
-from interactive_sdk.client.session import DefaultSession, Session
-from interactive_sdk.client.status import Status
